+++ pkg/application/ecosystemConfigUseCase_test.go
@@ -16,14 +16,17 @@ const (
 	testBlueprintID           = "blueprint1"
 )
 
-func TestEcosystemRegistryUseCase_ApplyConfig(t *testing.T) {
+func TestEcosystemConfigUseCase_ApplyConfig(t *testing.T) {
 	t.Run("success", func(t *testing.T) {
 		// given
 		blueprintRepoMock := newMockBlueprintSpecRepository(t)
 		doguConfigMock := newMockDoguConfigEntryRepository(t)
 		sensitiveDoguConfigMock := newMockSensitiveDoguConfigEntryRepository(t)
 		globalConfigMock := newMockGlobalConfigEntryRepository(t)
+		encryptionMock := newMockConfigEncryptionAdapter(t)
 
+		redmineDiffToEncrypt := getSensitiveDoguConfigEntryDiffForAction("key", "value", testSimpleDoguNameRedmine, domain.ConfigActionSetEncrypted)
+		casDiffToEncrypt := getSensitiveDoguConfigEntryDiffForAction("key", "value", testSimpleDoguNameCas, domain.ConfigActionSetEncrypted)
 		spec := &domain.BlueprintSpec{
 			StateDiff: domain.StateDiff{
 				DoguConfigDiffs: map[common.SimpleDoguName]domain.CombinedDoguConfigDiffs{
@@ -33,7 +36,7 @@ func TestEcosystemRegistryUseCase_ApplyConfig(t *testing.T) {
 							getRemoveDoguConfigEntryDiff("key", testSimpleDoguNameRedmine),
 						},
 						SensitiveDoguConfigDiff: []domain.SensitiveDoguConfigEntryDiff{
-							getSetEncryptedSensitiveDoguConfigEntryDiff("key", "value", testSimpleDoguNameRedmine),
+							redmineDiffToEncrypt,
 							getRemoveSensitiveDoguConfigEntryDiff("key", testSimpleDoguNameRedmine),
 						},
 					},
@@ -43,7 +46,7 @@ func TestEcosystemRegistryUseCase_ApplyConfig(t *testing.T) {
 							getRemoveDoguConfigEntryDiff("key", testSimpleDoguNameCas),
 						},
 						SensitiveDoguConfigDiff: []domain.SensitiveDoguConfigEntryDiff{
-							getSetEncryptedSensitiveDoguConfigEntryDiff("key", "value", testSimpleDoguNameCas),
+							casDiffToEncrypt,
 							getRemoveSensitiveDoguConfigEntryDiff("key", testSimpleDoguNameCas),
 						},
 					},
@@ -56,17 +59,21 @@ func TestEcosystemRegistryUseCase_ApplyConfig(t *testing.T) {
 		}
 
 		// Just check if the routine hits the repos. Check values in concrete test of methods.
-		doguConfigMock.EXPECT().SaveAll(testCtx, mock.Anything).Return(nil).Times(2)
-		doguConfigMock.EXPECT().DeleteAllByKeys(testCtx, mock.Anything).Return(nil).Times(2)
-		sensitiveDoguConfigMock.EXPECT().SaveAll(testCtx, mock.Anything).Return(nil).Times(2)
-		sensitiveDoguConfigMock.EXPECT().DeleteAllByKeys(testCtx, mock.Anything).Return(nil).Times(2)
+		doguConfigMock.EXPECT().SaveAll(testCtx, mock.Anything).Return(nil).Times(1)
+		doguConfigMock.EXPECT().DeleteAllByKeys(testCtx, mock.Anything).Return(nil).Times(1)
+		sensitiveDoguConfigMock.EXPECT().SaveAll(testCtx, mock.Anything).Return(nil).Times(1)
+		sensitiveDoguConfigMock.EXPECT().DeleteAllByKeys(testCtx, mock.Anything).Return(nil).Times(1)
 		globalConfigMock.EXPECT().SaveAll(testCtx, mock.Anything).Return(nil).Times(1)
 		globalConfigMock.EXPECT().DeleteAllByKeys(testCtx, mock.Anything).Return(nil).Times(1)
 
 		blueprintRepoMock.EXPECT().GetById(testCtx, testBlueprintID).Return(spec, nil)
 		blueprintRepoMock.EXPECT().Update(testCtx, mock.Anything).Return(nil).Times(2)
 
-		sut := EcosystemRegistryUseCase{blueprintRepository: blueprintRepoMock, doguConfigRepository: doguConfigMock, doguSensitiveConfigRepository: sensitiveDoguConfigMock, globalConfigRepository: globalConfigMock}
+		expectedSensitiveValueMapToEncrypt := map[common.SensitiveDoguConfigKey]common.SensitiveDoguConfigValue{redmineDiffToEncrypt.Key: common.SensitiveDoguConfigValue(redmineDiffToEncrypt.Expected.Value), casDiffToEncrypt.Key: common.SensitiveDoguConfigValue(casDiffToEncrypt.Expected.Value)}
+		encryptedDoguConfigValueEntries := map[common.SensitiveDoguConfigKey]common.EncryptedDoguConfigValue{redmineDiffToEncrypt.Key: common.EncryptedDoguConfigValue(""), casDiffToEncrypt.Key: common.EncryptedDoguConfigValue("")}
+		encryptionMock.EXPECT().EncryptAll(testCtx, expectedSensitiveValueMapToEncrypt).Return(encryptedDoguConfigValueEntries, nil).Times(1)
+
+		sut := EcosystemConfigUseCase{blueprintRepository: blueprintRepoMock, doguConfigRepository: doguConfigMock, doguSensitiveConfigRepository: sensitiveDoguConfigMock, globalConfigRepository: globalConfigMock, encryptionAdapter: encryptionMock}
 
 		// when
 		err := sut.ApplyConfig(testCtx, testBlueprintID)
@@ -81,7 +88,7 @@ func TestEcosystemRegistryUseCase_ApplyConfig(t *testing.T) {
 
 		blueprintRepoMock.EXPECT().GetById(testCtx, testBlueprintID).Return(nil, assert.AnError)
 
-		sut := EcosystemRegistryUseCase{blueprintRepository: blueprintRepoMock}
+		sut := EcosystemConfigUseCase{blueprintRepository: blueprintRepoMock}
 
 		// when
 		err := sut.ApplyConfig(testCtx, testBlueprintID)
@@ -106,7 +113,7 @@ func TestEcosystemRegistryUseCase_ApplyConfig(t *testing.T) {
 		blueprintRepoMock.EXPECT().GetById(testCtx, testBlueprintID).Return(spec, nil)
 		blueprintRepoMock.EXPECT().Update(testCtx, mock.Anything).Return(nil).Times(1)
 
-		sut := EcosystemRegistryUseCase{blueprintRepository: blueprintRepoMock}
+		sut := EcosystemConfigUseCase{blueprintRepository: blueprintRepoMock}
 
 		// when
 		err := sut.ApplyConfig(testCtx, testBlueprintID)
@@ -133,7 +140,7 @@ func TestEcosystemRegistryUseCase_ApplyConfig(t *testing.T) {
 		blueprintRepoMock.EXPECT().Update(testCtx, mock.Anything).Return(assert.AnError).Times(1)
 		blueprintRepoMock.EXPECT().Update(testCtx, mock.Anything).Return(nil).Times(1)
 
-		sut := EcosystemRegistryUseCase{blueprintRepository: blueprintRepoMock}
+		sut := EcosystemConfigUseCase{blueprintRepository: blueprintRepoMock}
 
 		// when
 		err := sut.ApplyConfig(testCtx, testBlueprintID)
@@ -149,7 +156,9 @@ func TestEcosystemRegistryUseCase_ApplyConfig(t *testing.T) {
 		doguConfigMock := newMockDoguConfigEntryRepository(t)
 		sensitiveDoguConfigMock := newMockSensitiveDoguConfigEntryRepository(t)
 		globalConfigMock := newMockGlobalConfigEntryRepository(t)
+		encryptionMock := newMockConfigEncryptionAdapter(t)
 
+		casDiffToEncrypt := getSensitiveDoguConfigEntryDiffForAction("key", "value", testSimpleDoguNameCas, domain.ConfigActionSetEncrypted)
 		spec := &domain.BlueprintSpec{
 			StateDiff: domain.StateDiff{
 				DoguConfigDiffs: map[common.SimpleDoguName]domain.CombinedDoguConfigDiffs{
@@ -158,7 +167,7 @@ func TestEcosystemRegistryUseCase_ApplyConfig(t *testing.T) {
 							getSetDoguConfigEntryDiff("key", "value", testSimpleDoguNameRedmine),
 						},
 						SensitiveDoguConfigDiff: []domain.SensitiveDoguConfigEntryDiff{
-							getSetEncryptedSensitiveDoguConfigEntryDiff("key", "value", testSimpleDoguNameRedmine),
+							getSensitiveDoguConfigEntryDiffForAction("key", "value", testSimpleDoguNameRedmine, domain.ConfigActionSetToEncrypt),
 						},
 					},
 					testSimpleDoguNameCas: {
@@ -166,7 +175,7 @@ func TestEcosystemRegistryUseCase_ApplyConfig(t *testing.T) {
 							getSetDoguConfigEntryDiff("key", "value", testSimpleDoguNameCas),
 						},
 						SensitiveDoguConfigDiff: []domain.SensitiveDoguConfigEntryDiff{
-							getSetEncryptedSensitiveDoguConfigEntryDiff("key", "value", testSimpleDoguNameCas),
+							casDiffToEncrypt,
 						},
 					},
 				},
@@ -177,14 +186,19 @@ func TestEcosystemRegistryUseCase_ApplyConfig(t *testing.T) {
 		}
 
 		// Just check if the routine hits the repos. Check values in concrete test of methods.
-		doguConfigMock.EXPECT().SaveAll(testCtx, mock.Anything).Return(assert.AnError).Times(2)
-		sensitiveDoguConfigMock.EXPECT().SaveAll(testCtx, mock.Anything).Return(assert.AnError).Times(2)
+		doguConfigMock.EXPECT().SaveAll(testCtx, mock.Anything).Return(assert.AnError).Times(1)
+		expectedCasMapToEncrypt := map[common.SensitiveDoguConfigKey]common.SensitiveDoguConfigValue{casDiffToEncrypt.Key: common.SensitiveDoguConfigValue(casDiffToEncrypt.Expected.Value)}
+		encryptedCasEntries := map[common.SensitiveDoguConfigKey]common.EncryptedDoguConfigValue{casDiffToEncrypt.Key: common.EncryptedDoguConfigValue("")}
+		encryptionMock.EXPECT().EncryptAll(testCtx, expectedCasMapToEncrypt).Return(encryptedCasEntries, nil).Times(1)
+
+		sensitiveDoguConfigMock.EXPECT().SaveAll(testCtx, mock.Anything).Return(assert.AnError).Times(1)
+		sensitiveDoguConfigMock.EXPECT().SaveAllForNotInstalledDogus(testCtx, mock.Anything).Return(assert.AnError).Times(1)
 		globalConfigMock.EXPECT().SaveAll(testCtx, mock.Anything).Return(assert.AnError).Times(1)
 
 		blueprintRepoMock.EXPECT().GetById(testCtx, testBlueprintID).Return(spec, nil)
 		blueprintRepoMock.EXPECT().Update(testCtx, mock.Anything).Return(nil).Times(2)
 
-		sut := EcosystemRegistryUseCase{blueprintRepository: blueprintRepoMock, doguConfigRepository: doguConfigMock, doguSensitiveConfigRepository: sensitiveDoguConfigMock, globalConfigRepository: globalConfigMock}
+		sut := EcosystemConfigUseCase{blueprintRepository: blueprintRepoMock, doguConfigRepository: doguConfigMock, doguSensitiveConfigRepository: sensitiveDoguConfigMock, globalConfigRepository: globalConfigMock, encryptionAdapter: encryptionMock}
 
 		// when
 		err := sut.ApplyConfig(testCtx, testBlueprintID)
@@ -193,18 +207,18 @@ func TestEcosystemRegistryUseCase_ApplyConfig(t *testing.T) {
 		require.NoError(t, err)
 		assert.Equal(t, spec.Status, domain.StatusPhaseApplyRegistryConfigFailed)
 		assert.Len(t, spec.Events, 2)
-		assert.Equal(t, spec.Events[1].Message(), "assert.AnError general error for testing\nassert.AnError general error for testing\nassert.AnError general error for testing\nassert.AnError general error for testing\nassert.AnError general error for testing")
+		assert.Equal(t, spec.Events[1].Message(), "assert.AnError general error for testing\nassert.AnError general error for testing\nassert.AnError general error for testing\nassert.AnError general error for testing")
 	})
 }
 
-func TestEcosystemRegistryUseCase_applyDoguConfigDiffs(t *testing.T) {
+func TestEcosystemConfigUseCase_applyDoguConfigDiffs(t *testing.T) {
 	t.Run("should save diffs with action set", func(t *testing.T) {
 		// given
 		doguConfigMock := newMockDoguConfigEntryRepository(t)
-		sut := NewEcosystemRegistryUseCase(nil, doguConfigMock, nil, nil)
+		sut := NewEcosystemConfigUseCase(nil, doguConfigMock, nil, nil, nil)
 		diff1 := getSetDoguConfigEntryDiff("/key", "value", testSimpleDoguNameRedmine)
 		diff2 := getSetDoguConfigEntryDiff("/key1", "value1", testSimpleDoguNameRedmine)
-		diffs := domain.DoguConfigDiffs{diff1, diff2}
+		byAction := map[domain.ConfigAction]domain.DoguConfigDiffs{domain.ConfigActionSet: {diff1, diff2}}
 
 		expectedEntry1 := &ecosystem.DoguConfigEntry{
 			Key:   common.DoguConfigKey{DoguName: testSimpleDoguNameRedmine, Key: diff1.Key.Key},
@@ -218,7 +232,7 @@ func TestEcosystemRegistryUseCase_applyDoguConfigDiffs(t *testing.T) {
 		doguConfigMock.EXPECT().SaveAll(testCtx, []*ecosystem.DoguConfigEntry{expectedEntry1, expectedEntry2}).Return(nil).Times(1)
 
 		// when
-		err := sut.applyDoguConfigDiffs(testCtx, testSimpleDoguNameRedmine, diffs)
+		err := sut.applyDoguConfigDiffs(testCtx, byAction)
 
 		// then
 		require.NoError(t, err)
@@ -227,10 +241,10 @@ func TestEcosystemRegistryUseCase_applyDoguConfigDiffs(t *testing.T) {
 	t.Run("should delete diffs with action remove", func(t *testing.T) {
 		// given
 		doguConfigMock := newMockDoguConfigEntryRepository(t)
-		sut := NewEcosystemRegistryUseCase(nil, doguConfigMock, nil, nil)
+		sut := NewEcosystemConfigUseCase(nil, doguConfigMock, nil, nil, nil)
 		diff1 := getRemoveDoguConfigEntryDiff("/key", testSimpleDoguNameRedmine)
 		diff2 := getRemoveDoguConfigEntryDiff("/key1", testSimpleDoguNameRedmine)
-		diffs := domain.DoguConfigDiffs{diff1, diff2}
+		byAction := map[domain.ConfigAction]domain.DoguConfigDiffs{domain.ConfigActionRemove: {diff1, diff2}}
 
 		expectedKey1 := common.DoguConfigKey{DoguName: testSimpleDoguNameRedmine, Key: diff1.Key.Key}
 		expectedKey2 := common.DoguConfigKey{DoguName: testSimpleDoguNameRedmine, Key: diff2.Key.Key}
@@ -238,7 +252,7 @@ func TestEcosystemRegistryUseCase_applyDoguConfigDiffs(t *testing.T) {
 		doguConfigMock.EXPECT().DeleteAllByKeys(testCtx, []common.DoguConfigKey{expectedKey1, expectedKey2}).Return(nil).Times(1)
 
 		// when
-		err := sut.applyDoguConfigDiffs(testCtx, testSimpleDoguNameRedmine, diffs)
+		err := sut.applyDoguConfigDiffs(testCtx, byAction)
 
 		// then
 		require.NoError(t, err)
@@ -247,47 +261,28 @@ func TestEcosystemRegistryUseCase_applyDoguConfigDiffs(t *testing.T) {
 	t.Run("should return nil on action none", func(t *testing.T) {
 		// given
 		doguConfigMock := newMockDoguConfigEntryRepository(t)
-		sut := NewEcosystemRegistryUseCase(nil, doguConfigMock, nil, nil)
+		sut := NewEcosystemConfigUseCase(nil, doguConfigMock, nil, nil, nil)
 		diff1 := domain.DoguConfigEntryDiff{
 			NeededAction: domain.ConfigActionNone,
 		}
-
-		diffs := domain.DoguConfigDiffs{diff1}
+		byAction := map[domain.ConfigAction]domain.DoguConfigDiffs{domain.ConfigActionNone: {diff1}}
 
 		// when
-		err := sut.applyDoguConfigDiffs(testCtx, testSimpleDoguNameRedmine, diffs)
+		err := sut.applyDoguConfigDiffs(testCtx, byAction)
 
 		// then
 		require.NoError(t, err)
 	})
-
-	t.Run("should return error on unknown action", func(t *testing.T) {
-		// given
-		sut := NewEcosystemRegistryUseCase(nil, newMockDoguConfigEntryRepository(t), nil, nil)
-		diff1 := domain.DoguConfigEntryDiff{
-			Key:          common.DoguConfigKey{Key: "key"},
-			NeededAction: "unknown",
-		}
-
-		diffs := domain.DoguConfigDiffs{diff1}
-
-		// when
-		err := sut.applyDoguConfigDiffs(testCtx, testSimpleDoguNameRedmine, diffs)
-
-		// then
-		require.Error(t, err)
-		assert.ErrorContains(t, err, "cannot perform unknown action \"unknown\" for dogu \"redmine\" with key \"key\"")
-	})
 }
 
-func TestEcosystemRegistryUseCase_applyGlobalConfigDiffs(t *testing.T) {
+func TestEcosystemConfigUseCase_applyGlobalConfigDiffs(t *testing.T) {
 	t.Run("should save diffs with action set", func(t *testing.T) {
 		// given
 		globalConfigMock := newMockGlobalConfigEntryRepository(t)
-		sut := NewEcosystemRegistryUseCase(nil, nil, nil, globalConfigMock)
+		sut := NewEcosystemConfigUseCase(nil, nil, nil, globalConfigMock, nil)
 		diff1 := getSetGlobalConfigEntryDiff("/key", "value")
 		diff2 := getSetGlobalConfigEntryDiff("/key1", "value1")
-		diffs := domain.GlobalConfigDiffs{diff1, diff2}
+		byAction := map[domain.ConfigAction][]domain.GlobalConfigEntryDiff{domain.ConfigActionSet: {diff1, diff2}}
 
 		expectedEntry1 := &ecosystem.GlobalConfigEntry{
 			Key:   diff1.Key,
@@ -301,7 +296,7 @@ func TestEcosystemRegistryUseCase_applyGlobalConfigDiffs(t *testing.T) {
 		globalConfigMock.EXPECT().SaveAll(testCtx, []*ecosystem.GlobalConfigEntry{expectedEntry1, expectedEntry2}).Return(nil).Times(1)
 
 		// when
-		err := sut.applyGlobalConfigDiffs(testCtx, diffs)
+		err := sut.applyGlobalConfigDiffs(testCtx, byAction)
 
 		// then
 		require.NoError(t, err)
@@ -310,15 +305,15 @@ func TestEcosystemRegistryUseCase_applyGlobalConfigDiffs(t *testing.T) {
 	t.Run("should delete diffs with action remove", func(t *testing.T) {
 		// given
 		globalConfigMock := newMockGlobalConfigEntryRepository(t)
-		sut := NewEcosystemRegistryUseCase(nil, nil, nil, globalConfigMock)
+		sut := NewEcosystemConfigUseCase(nil, nil, nil, globalConfigMock, nil)
 		diff1 := getRemoveGlobalConfigEntryDiff("/key")
 		diff2 := getRemoveGlobalConfigEntryDiff("/key1")
-		diffs := domain.GlobalConfigDiffs{diff1, diff2}
+		byAction := map[domain.ConfigAction][]domain.GlobalConfigEntryDiff{domain.ConfigActionRemove: {diff1, diff2}}
 
 		globalConfigMock.EXPECT().DeleteAllByKeys(testCtx, []common.GlobalConfigKey{diff1.Key, diff2.Key}).Return(nil).Times(1)
 
 		// when
-		err := sut.applyGlobalConfigDiffs(testCtx, diffs)
+		err := sut.applyGlobalConfigDiffs(testCtx, byAction)
 
 		// then
 		require.NoError(t, err)
@@ -326,48 +321,59 @@ func TestEcosystemRegistryUseCase_applyGlobalConfigDiffs(t *testing.T) {
 
 	t.Run("should return nil on action none", func(t *testing.T) {
 		// given
-		sut := NewEcosystemRegistryUseCase(nil, nil, nil, newMockGlobalConfigEntryRepository(t))
+		sut := NewEcosystemConfigUseCase(nil, nil, nil, newMockGlobalConfigEntryRepository(t), nil)
 		diff1 := domain.GlobalConfigEntryDiff{
 			NeededAction: domain.ConfigActionNone,
 		}
-
-		diffs := domain.GlobalConfigDiffs{diff1}
+		byAction := map[domain.ConfigAction][]domain.GlobalConfigEntryDiff{domain.ConfigActionNone: {diff1}}
 
 		// when
-		err := sut.applyGlobalConfigDiffs(testCtx, diffs)
+		err := sut.applyGlobalConfigDiffs(testCtx, byAction)
+
+		// then
+		require.NoError(t, err)
+	})
+}
+
+func TestEcosystemConfigUseCase_applySensitiveDoguConfigDiffs(t *testing.T) {
+	t.Run("should save diffs with action setEncrypted", func(t *testing.T) {
+		// given
+		sensitiveDoguConfigMock := newMockSensitiveDoguConfigEntryRepository(t)
+		encryptionMock := newMockConfigEncryptionAdapter(t)
+		sut := NewEcosystemConfigUseCase(nil, nil, sensitiveDoguConfigMock, nil, encryptionMock)
+		diff1 := getSensitiveDoguConfigEntryDiffForAction("key", "value", testSimpleDoguNameRedmine, domain.ConfigActionSetEncrypted)
+		diff2 := getSensitiveDoguConfigEntryDiffForAction("key1", "value1", testSimpleDoguNameRedmine, domain.ConfigActionSetEncrypted)
+		byAction := map[domain.ConfigAction]domain.SensitiveDoguConfigDiffs{domain.ConfigActionSetEncrypted: {diff1, diff2}}
+
+		expectedRedmineMapToEncrypt := map[common.SensitiveDoguConfigKey]common.SensitiveDoguConfigValue{diff1.Key: common.SensitiveDoguConfigValue(diff1.Expected.Value), diff2.Key: common.SensitiveDoguConfigValue(diff2.Expected.Value)}
+		encryptedRedmineEntries := map[common.SensitiveDoguConfigKey]common.EncryptedDoguConfigValue{diff1.Key: common.EncryptedDoguConfigValue("encrypted_value"), diff2.Key: common.EncryptedDoguConfigValue("encrypted_value1")}
+		encryptionMock.EXPECT().EncryptAll(testCtx, expectedRedmineMapToEncrypt).Return(encryptedRedmineEntries, nil)
+
+		expectedEntry1 := &ecosystem.SensitiveDoguConfigEntry{
+			Key:   common.SensitiveDoguConfigKey{DoguConfigKey: common.DoguConfigKey{DoguName: testSimpleDoguNameRedmine, Key: diff1.Key.Key}},
+			Value: common.EncryptedDoguConfigValue("encrypted_value"),
+		}
+		expectedEntry2 := &ecosystem.SensitiveDoguConfigEntry{
+			Key:   common.SensitiveDoguConfigKey{DoguConfigKey: common.DoguConfigKey{DoguName: testSimpleDoguNameRedmine, Key: diff2.Key.Key}},
+			Value: common.EncryptedDoguConfigValue("encrypted_value1"),
+		}
+
+		sensitiveDoguConfigMock.EXPECT().SaveAll(testCtx, []*ecosystem.SensitiveDoguConfigEntry{expectedEntry1, expectedEntry2}).Return(nil).Times(1)
+
+		// when
+		err := sut.applySensitiveDoguConfigDiffs(testCtx, byAction)
 
 		// then
 		require.NoError(t, err)
 	})
 
-	t.Run("should return error on unknown action", func(t *testing.T) {
-		// given
-		globalConfigMock := newMockGlobalConfigEntryRepository(t)
-		sut := NewEcosystemRegistryUseCase(nil, nil, nil, globalConfigMock)
-		diff1 := domain.GlobalConfigEntryDiff{
-			Key:          "key",
-			NeededAction: "unknown",
-		}
-
-		diffs := domain.GlobalConfigDiffs{diff1}
-
-		// when
-		err := sut.applyGlobalConfigDiffs(testCtx, diffs)
-
-		// then
-		require.Error(t, err)
-		assert.ErrorContains(t, err, "cannot perform unknown action \"unknown\" for global config with key \"key\"")
-	})
-}
-
-func TestEcosystemRegistryUseCase_applySensitiveDoguConfigDiffs(t *testing.T) {
-	t.Run("should save diffs with action set", func(t *testing.T) {
+	t.Run("should save diffs with action setToEncrypt", func(t *testing.T) {
 		// given
 		sensitiveDoguConfigMock := newMockSensitiveDoguConfigEntryRepository(t)
-		sut := NewEcosystemRegistryUseCase(nil, nil, sensitiveDoguConfigMock, nil)
-		diff1 := getSetEncryptedSensitiveDoguConfigEntryDiff("key", "value", testSimpleDoguNameRedmine)
-		diff2 := getSetEncryptedSensitiveDoguConfigEntryDiff("key1", "value1", testSimpleDoguNameRedmine)
-		diffs := domain.SensitiveDoguConfigDiffs{diff1, diff2}
+		sut := NewEcosystemConfigUseCase(nil, nil, sensitiveDoguConfigMock, nil, nil)
+		diff1 := getSensitiveDoguConfigEntryDiffForAction("key", "value", testSimpleDoguNameRedmine, domain.ConfigActionSetToEncrypt)
+		diff2 := getSensitiveDoguConfigEntryDiffForAction("key1", "value1", testSimpleDoguNameRedmine, domain.ConfigActionSetToEncrypt)
+		byAction := map[domain.ConfigAction]domain.SensitiveDoguConfigDiffs{domain.ConfigActionSetToEncrypt: {diff1, diff2}}
 
 		expectedEntry1 := &ecosystem.SensitiveDoguConfigEntry{
 			Key:   common.SensitiveDoguConfigKey{DoguConfigKey: common.DoguConfigKey{DoguName: testSimpleDoguNameRedmine, Key: diff1.Key.Key}},
@@ -378,10 +384,10 @@ func TestEcosystemRegistryUseCase_applySensitiveDoguConfigDiffs(t *testing.T) {
 			Value: common.EncryptedDoguConfigValue(diff2.Expected.Value),
 		}
 
-		sensitiveDoguConfigMock.EXPECT().SaveAll(testCtx, []*ecosystem.SensitiveDoguConfigEntry{expectedEntry1, expectedEntry2}).Return(nil).Times(1)
+		sensitiveDoguConfigMock.EXPECT().SaveAllForNotInstalledDogus(testCtx, []*ecosystem.SensitiveDoguConfigEntry{expectedEntry1, expectedEntry2}).Return(nil).Times(1)
 
 		// when
-		err := sut.applySensitiveDoguConfigDiffs(testCtx, testSimpleDoguNameRedmine, diffs)
+		err := sut.applySensitiveDoguConfigDiffs(testCtx, byAction)
 
 		// then
 		require.NoError(t, err)
@@ -390,10 +396,10 @@ func TestEcosystemRegistryUseCase_applySensitiveDoguConfigDiffs(t *testing.T) {
 	t.Run("should delete diffs with action remove", func(t *testing.T) {
 		// given
 		sensitiveDoguConfigMock := newMockSensitiveDoguConfigEntryRepository(t)
-		sut := NewEcosystemRegistryUseCase(nil, nil, sensitiveDoguConfigMock, nil)
+		sut := NewEcosystemConfigUseCase(nil, nil, sensitiveDoguConfigMock, nil, nil)
 		diff1 := getRemoveSensitiveDoguConfigEntryDiff("key", testSimpleDoguNameRedmine)
 		diff2 := getRemoveSensitiveDoguConfigEntryDiff("key", testSimpleDoguNameRedmine)
-		diffs := domain.SensitiveDoguConfigDiffs{diff1, diff2}
+		byAction := map[domain.ConfigAction]domain.SensitiveDoguConfigDiffs{domain.ConfigActionRemove: {diff1, diff2}}
 
 		expectedKey1 := common.SensitiveDoguConfigKey{DoguConfigKey: common.DoguConfigKey{DoguName: testSimpleDoguNameRedmine, Key: diff1.Key.Key}}
 		expectedKey2 := common.SensitiveDoguConfigKey{DoguConfigKey: common.DoguConfigKey{DoguName: testSimpleDoguNameRedmine, Key: diff2.Key.Key}}
@@ -401,49 +407,93 @@ func TestEcosystemRegistryUseCase_applySensitiveDoguConfigDiffs(t *testing.T) {
 		sensitiveDoguConfigMock.EXPECT().DeleteAllByKeys(testCtx, []common.SensitiveDoguConfigKey{expectedKey1, expectedKey2}).Return(nil).Times(1)
 
 		// when
-		err := sut.applySensitiveDoguConfigDiffs(testCtx, testSimpleDoguNameRedmine, diffs)
+		err := sut.applySensitiveDoguConfigDiffs(testCtx, byAction)
 
 		// then
 		require.NoError(t, err)
+	})
+
+	t.Run("should return error on encryption error", func(t *testing.T) {
+		// given
+		sensitiveDoguConfigMock := newMockSensitiveDoguConfigEntryRepository(t)
+		encryptionMock := newMockConfigEncryptionAdapter(t)
+		encryptionMock.EXPECT().EncryptAll(testCtx, mock.Anything).Return(nil, assert.AnError)
+
+		diff := getSensitiveDoguConfigEntryDiffForAction("key", "value", testSimpleDoguNameRedmine, domain.ConfigActionSetEncrypted)
+		byAction := map[domain.ConfigAction]domain.SensitiveDoguConfigDiffs{domain.ConfigActionSetEncrypted: {diff}}
+
+		sut := EcosystemConfigUseCase{encryptionAdapter: encryptionMock, doguSensitiveConfigRepository: sensitiveDoguConfigMock}
+
+		// when
+		err := sut.applySensitiveDoguConfigDiffs(testCtx, byAction)
+
+		// then
+		require.Error(t, err)
+		assert.ErrorIs(t, err, assert.AnError)
+	})
+
+	t.Run("should return error if map with encrypted values is nil", func(t *testing.T) {
+		// given
+		doguConfigMock := newMockDoguConfigEntryRepository(t)
+		sensitiveDoguConfigMock := newMockSensitiveDoguConfigEntryRepository(t)
+		globalConfigMock := newMockGlobalConfigEntryRepository(t)
+		encryptionMock := newMockConfigEncryptionAdapter(t)
+		encryptionMock.EXPECT().EncryptAll(testCtx, mock.Anything).Return(nil, nil)
+
+		diff := getSensitiveDoguConfigEntryDiffForAction("key", "value", testSimpleDoguNameRedmine, domain.ConfigActionSetEncrypted)
+		byAction := map[domain.ConfigAction]domain.SensitiveDoguConfigDiffs{domain.ConfigActionSetEncrypted: {diff}}
+
+		sut := EcosystemConfigUseCase{doguConfigRepository: doguConfigMock, doguSensitiveConfigRepository: sensitiveDoguConfigMock, globalConfigRepository: globalConfigMock, encryptionAdapter: encryptionMock}
+
+		// when
+		err := sut.applySensitiveDoguConfigDiffs(testCtx, byAction)
+
+		// then
+		require.Error(t, err)
+		assert.ErrorContains(t, err, "encrypted entry value map is nil: sensitive dogu config error")
+	})
+
+	t.Run("should return error if map contains not required sensitive dogu config key", func(t *testing.T) {
+		// given
+		doguConfigMock := newMockDoguConfigEntryRepository(t)
+		sensitiveDoguConfigMock := newMockSensitiveDoguConfigEntryRepository(t)
+		globalConfigMock := newMockGlobalConfigEntryRepository(t)
+		encryptionMock := newMockConfigEncryptionAdapter(t)
+
+		diff := getSensitiveDoguConfigEntryDiffForAction("key", "value", testSimpleDoguNameRedmine, domain.ConfigActionSetEncrypted)
+		byAction := map[domain.ConfigAction]domain.SensitiveDoguConfigDiffs{domain.ConfigActionSetEncrypted: {diff}}
+		expectedRedmineMapToEncrypt := map[common.SensitiveDoguConfigKey]common.SensitiveDoguConfigValue{diff.Key: common.SensitiveDoguConfigValue(diff.Expected.Value)}
+		otherKey := common.SensitiveDoguConfigKey{}
+		encryptedRedmineEntries := map[common.SensitiveDoguConfigKey]common.EncryptedDoguConfigValue{otherKey: common.EncryptedDoguConfigValue("encrypted_value")}
+		encryptionMock.EXPECT().EncryptAll(testCtx, expectedRedmineMapToEncrypt).Return(encryptedRedmineEntries, nil)
+
+		sut := EcosystemConfigUseCase{doguConfigRepository: doguConfigMock, doguSensitiveConfigRepository: sensitiveDoguConfigMock, globalConfigRepository: globalConfigMock, encryptionAdapter: encryptionMock}
+
+		// when
+		err := sut.applySensitiveDoguConfigDiffs(testCtx, byAction)
+
+		// then
+		require.Error(t, err)
+		assert.ErrorContains(t, err, "did not find encrypted value for key key: sensitive dogu config error")
 	})
 
 	t.Run("should return nil on action none", func(t *testing.T) {
 		// given
-		sut := NewEcosystemRegistryUseCase(nil, nil, newMockSensitiveDoguConfigEntryRepository(t), nil)
+		sut := NewEcosystemConfigUseCase(nil, nil, newMockSensitiveDoguConfigEntryRepository(t), nil, nil)
 		diff1 := domain.SensitiveDoguConfigEntryDiff{
 			NeededAction: domain.ConfigActionNone,
 		}
-
-		diffs := domain.SensitiveDoguConfigDiffs{diff1}
+		byAction := map[domain.ConfigAction]domain.SensitiveDoguConfigDiffs{domain.ConfigActionNone: {diff1}}
 
 		// when
-		err := sut.applySensitiveDoguConfigDiffs(testCtx, testSimpleDoguNameRedmine, diffs)
+		err := sut.applySensitiveDoguConfigDiffs(testCtx, byAction)
 
 		// then
 		require.NoError(t, err)
 	})
-
-	t.Run("should return error on unknown action", func(t *testing.T) {
-		// given
-		sensitiveDoguConfigMock := newMockSensitiveDoguConfigEntryRepository(t)
-		sut := NewEcosystemRegistryUseCase(nil, nil, sensitiveDoguConfigMock, nil)
-		diff1 := domain.SensitiveDoguConfigEntryDiff{
-			Key:          common.SensitiveDoguConfigKey{DoguConfigKey: common.DoguConfigKey{Key: "key"}},
-			NeededAction: "unknown",
-		}
-
-		diffs := domain.SensitiveDoguConfigDiffs{diff1}
-
-		// when
-		err := sut.applySensitiveDoguConfigDiffs(testCtx, testSimpleDoguNameRedmine, diffs)
-
-		// then
-		require.Error(t, err)
-		assert.ErrorContains(t, err, "cannot perform unknown action \"unknown\" for dogu \"redmine\" with key \"key\"")
-	})
 }
 
-func TestEcosystemRegistryUseCase_markConfigApplied(t *testing.T) {
+func TestEcosystemConfigUseCase_markConfigApplied(t *testing.T) {
 	t.Run("should set applied status and event", func(t *testing.T) {
 		// given
 		spec := &domain.BlueprintSpec{}
@@ -454,7 +504,7 @@ func TestEcosystemRegistryUseCase_markConfigApplied(t *testing.T) {
 
 		blueprintRepoMock.EXPECT().Update(testCtx, expectedSpec).Return(nil)
 
-		sut := EcosystemRegistryUseCase{blueprintRepository: blueprintRepoMock}
+		sut := EcosystemConfigUseCase{blueprintRepository: blueprintRepoMock}
 
 		// when
 		err := sut.markConfigApplied(testCtx, spec)
@@ -473,7 +523,7 @@ func TestEcosystemRegistryUseCase_markConfigApplied(t *testing.T) {
 
 		blueprintRepoMock.EXPECT().Update(testCtx, expectedSpec).Return(assert.AnError)
 
-		sut := EcosystemRegistryUseCase{blueprintRepository: blueprintRepoMock}
+		sut := EcosystemConfigUseCase{blueprintRepository: blueprintRepoMock}
 
 		// when
 		err := sut.markConfigApplied(testCtx, spec)
@@ -485,7 +535,7 @@ func TestEcosystemRegistryUseCase_markConfigApplied(t *testing.T) {
 	})
 }
 
-func TestEcosystemRegistryUseCase_markApplyConfigStart(t *testing.T) {
+func TestEcosystemConfigUseCase_markApplyConfigStart(t *testing.T) {
 	t.Run("should set status and event apply config", func(t *testing.T) {
 		// given
 		spec := &domain.BlueprintSpec{}
@@ -496,7 +546,7 @@ func TestEcosystemRegistryUseCase_markApplyConfigStart(t *testing.T) {
 
 		blueprintRepoMock.EXPECT().Update(testCtx, expectedSpec).Return(nil)
 
-		sut := EcosystemRegistryUseCase{blueprintRepository: blueprintRepoMock}
+		sut := EcosystemConfigUseCase{blueprintRepository: blueprintRepoMock}
 
 		// when
 		err := sut.markApplyConfigStart(testCtx, spec)
@@ -515,7 +565,7 @@ func TestEcosystemRegistryUseCase_markApplyConfigStart(t *testing.T) {
 
 		blueprintRepoMock.EXPECT().Update(testCtx, expectedSpec).Return(assert.AnError)
 
-		sut := EcosystemRegistryUseCase{blueprintRepository: blueprintRepoMock}
+		sut := EcosystemConfigUseCase{blueprintRepository: blueprintRepoMock}
 
 		// when
 		err := sut.markApplyConfigStart(testCtx, spec)
@@ -527,7 +577,7 @@ func TestEcosystemRegistryUseCase_markApplyConfigStart(t *testing.T) {
 	})
 }
 
-func TestEcosystemRegistryUseCase_handleFailedApplyRegistryConfig(t *testing.T) {
+func TestEcosystemConfigUseCase_handleFailedApplyRegistryConfig(t *testing.T) {
 	t.Run("should set applied status and event", func(t *testing.T) {
 		// given
 		spec := &domain.BlueprintSpec{}
@@ -535,7 +585,7 @@ func TestEcosystemRegistryUseCase_handleFailedApplyRegistryConfig(t *testing.T) 
 
 		blueprintRepoMock.EXPECT().Update(testCtx, mock.IsType(&domain.BlueprintSpec{})).Return(nil)
 
-		sut := EcosystemRegistryUseCase{blueprintRepository: blueprintRepoMock}
+		sut := EcosystemConfigUseCase{blueprintRepository: blueprintRepoMock}
 
 		// when
 		err := sut.handleFailedApplyRegistryConfig(testCtx, spec, assert.AnError)
@@ -553,7 +603,7 @@ func TestEcosystemRegistryUseCase_handleFailedApplyRegistryConfig(t *testing.T) 
 
 		blueprintRepoMock.EXPECT().Update(testCtx, mock.IsType(&domain.BlueprintSpec{})).Return(assert.AnError)
 
-		sut := EcosystemRegistryUseCase{blueprintRepository: blueprintRepoMock}
+		sut := EcosystemConfigUseCase{blueprintRepository: blueprintRepoMock}
 
 		// when
 		err := sut.handleFailedApplyRegistryConfig(testCtx, spec, assert.AnError)
@@ -565,7 +615,7 @@ func TestEcosystemRegistryUseCase_handleFailedApplyRegistryConfig(t *testing.T) 
 	})
 }
 
-func TestNewEcosystemRegistryUseCase(t *testing.T) {
+func TestNewEcosystemConfigUseCase(t *testing.T) {
 	t.Run("success", func(t *testing.T) {
 		// given
 		blueprintRepoMock := newMockBlueprintSpecRepository(t)
@@ -574,7 +624,7 @@ func TestNewEcosystemRegistryUseCase(t *testing.T) {
 		globalConfigMock := newMockGlobalConfigEntryRepository(t)
 
 		// when
-		useCase := NewEcosystemRegistryUseCase(blueprintRepoMock, doguConfigMock, sensitiveDoguConfigMock, globalConfigMock)
+		useCase := NewEcosystemConfigUseCase(blueprintRepoMock, doguConfigMock, sensitiveDoguConfigMock, globalConfigMock, nil)
 
 		// then
 		assert.Equal(t, blueprintRepoMock, useCase.blueprintRepository)
@@ -607,7 +657,7 @@ func getRemoveDoguConfigEntryDiff(key string, doguName common.SimpleDoguName) do
 	}
 }
 
-func getSetEncryptedSensitiveDoguConfigEntryDiff(key, value string, doguName common.SimpleDoguName) domain.SensitiveDoguConfigEntryDiff {
+func getSensitiveDoguConfigEntryDiffForAction(key, value string, doguName common.SimpleDoguName, action domain.ConfigAction) domain.SensitiveDoguConfigEntryDiff {
 	return domain.SensitiveDoguConfigEntryDiff{
 		Key: common.SensitiveDoguConfigKey{
 			DoguConfigKey: common.DoguConfigKey{
@@ -618,7 +668,7 @@ func getSetEncryptedSensitiveDoguConfigEntryDiff(key, value string, doguName com
 		Expected: domain.DoguConfigValueState{
 			Value: value,
 		},
-		NeededAction: domain.ConfigActionSetEncrypted,
+		NeededAction: action,
 	}
 }