+++ pkg/domain/blueprintSpec_test.go
@@ -4,6 +4,7 @@ import (
 	"errors"
 	"fmt"
 	"github.com/cloudogu/k8s-blueprint-operator/pkg/domain/common"
+	"golang.org/x/exp/maps"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -707,6 +708,66 @@ func TestBlueprintSpec_MarkBlueprintApplied(t *testing.T) {
 		Status: StatusPhaseBlueprintApplied,
 		Events: []Event{BlueprintAppliedEvent{}},
 	})
+}
+
+func TestBlueprintSpec_CensorSensitiveData(t *testing.T) {
+	// given
+	ldapLoggingKey := common.SensitiveDoguConfigKey{DoguConfigKey: common.DoguConfigKey{DoguName: "ldap", Key: "logging/root"}}
+	spec := &BlueprintSpec{
+		Blueprint: Blueprint{
+			Config: Config{
+				Dogus: map[common.SimpleDoguName]CombinedDoguConfig{
+					"ldap": {
+						DoguName: "ldap",
+						SensitiveConfig: SensitiveDoguConfig{
+							Present: map[common.SensitiveDoguConfigKey]common.SensitiveDoguConfigValue{
+								ldapLoggingKey: "ERROR",
+							},
+						},
+					},
+				},
+			},
+		},
+		EffectiveBlueprint: EffectiveBlueprint{
+			Config: Config{
+				Dogus: map[common.SimpleDoguName]CombinedDoguConfig{
+					"ldap": {
+						DoguName: "ldap",
+						SensitiveConfig: SensitiveDoguConfig{
+							Present: map[common.SensitiveDoguConfigKey]common.SensitiveDoguConfigValue{
+								ldapLoggingKey: "ERROR",
+							},
+						},
+					},
+				},
+			},
+		},
+		StateDiff: StateDiff{
+			DoguConfigDiffs: map[common.SimpleDoguName]CombinedDoguConfigDiffs{
+				"ldapDiff": {SensitiveDoguConfigDiff: []SensitiveDoguConfigEntryDiff{{
+					Actual:   DoguConfigValueState{Value: "Test1"},
+					Expected: DoguConfigValueState{Value: "Test2"},
+				}}},
+			},
+		},
+	}
+	// when
+	spec.CensorSensitiveData()
+
+	// then
+	require.Len(t, spec.Blueprint.Config.Dogus, 1)
+	assert.Contains(t, maps.Keys(spec.Blueprint.Config.Dogus), common.SimpleDoguName("ldap"))
+	assert.Equal(t, censorValue, string(spec.Blueprint.Config.Dogus["ldap"].SensitiveConfig.Present[ldapLoggingKey]))
+
+	require.Len(t, spec.EffectiveBlueprint.Config.Dogus, 1)
+	assert.Contains(t, maps.Keys(spec.EffectiveBlueprint.Config.Dogus), common.SimpleDoguName("ldap"))
+	assert.Equal(t, censorValue, string(spec.EffectiveBlueprint.Config.Dogus["ldap"].SensitiveConfig.Present[ldapLoggingKey]))
+
+	require.Len(t, spec.StateDiff.DoguConfigDiffs, 1)
+	assert.Contains(t, maps.Keys(spec.StateDiff.DoguConfigDiffs), common.SimpleDoguName("ldapDiff"))
+	require.Len(t, spec.StateDiff.DoguConfigDiffs["ldapDiff"].SensitiveDoguConfigDiff, 1)
+	assert.Equal(t, censorValue, spec.StateDiff.DoguConfigDiffs["ldapDiff"].SensitiveDoguConfigDiff[0].Actual.Value)
+	assert.Equal(t, censorValue, spec.StateDiff.DoguConfigDiffs["ldapDiff"].SensitiveDoguConfigDiff[0].Expected.Value)
 }
 
 func TestBlueprintSpec_CompletePostProcessing(t *testing.T) {