+++ CHANGELOG.md
@@ -37,3 +37,5 @@ Config-specific:
 - [#48] Save config diff in cluster
 - [#39] Encrypt sensitive data
 - [#36] Set registry configuration for dogu and global config.
+- [#38] Censor all sensitive configuration data after applying the blueprint
+- [#45] Set registry configuration for encrypted values.